Move Attachment's children component close to it: Audio, CollapsibleQuote, Image, Reply, Video
