+++ app/containers/message/Components/Attachments/CollapsibleQuote/index.tsx
@@ -3,16 +3,16 @@ import { dequal } from 'dequal';
 import React, { useContext, useState } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 
-import { themes } from '../../../../lib/constants';
-import { IAttachment } from '../../../../definitions/IAttachment';
-import { TGetCustomEmoji } from '../../../../definitions/IEmoji';
-import { CustomIcon } from '../../../CustomIcon';
-import { useTheme } from '../../../../theme';
-import sharedStyles from '../../../../views/Styles';
-import Markdown from '../../../markdown';
-import MessageContext from '../../Context';
-import Touchable from '../../Touchable';
-import { BUTTON_HIT_SLOP } from '../../utils';
+import { themes } from '../../../../../lib/constants';
+import { IAttachment } from '../../../../../definitions/IAttachment';
+import { TGetCustomEmoji } from '../../../../../definitions/IEmoji';
+import { CustomIcon } from '../../../../CustomIcon';
+import { useTheme } from '../../../../../theme';
+import sharedStyles from '../../../../../views/Styles';
+import Markdown from '../../../../markdown';
+import MessageContext from '../../../Context';
+import Touchable from '../../../Touchable';
+import { BUTTON_HIT_SLOP } from '../../../utils';
 
 const styles = StyleSheet.create({
 	button: {
@@ -156,8 +156,7 @@ const CollapsibleQuote = React.memo(
 						}
 					]}
 					background={Touchable.Ripple(themes[theme].surfaceNeutral)}
-					hitSlop={BUTTON_HIT_SLOP}
-				>
+					hitSlop={BUTTON_HIT_SLOP}>
 					<View style={styles.touchableContainer}>
 						<View style={styles.attachmentContainer}>
 							<View style={styles.authorContainer}>
