+++ app/containers/message/Components/Attachments/Reply.tsx
@@ -4,19 +4,19 @@ import React, { useContext, useState } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import FastImage from 'react-native-fast-image';
 
-import { IAttachment, TGetCustomEmoji } from '../../definitions';
-import { themes } from '../../lib/constants';
-import { fileDownloadAndPreview } from '../../lib/methods/helpers';
-import { formatAttachmentUrl } from '../../lib/methods/helpers/formatAttachmentUrl';
-import openLink from '../../lib/methods/helpers/openLink';
-import { TSupportedThemes, useTheme } from '../../theme';
-import sharedStyles from '../../views/Styles';
-import RCActivityIndicator from '../ActivityIndicator';
-import Markdown from '../markdown';
+import { IAttachment, TGetCustomEmoji } from '../../../../definitions';
+import { themes } from '../../../../lib/constants';
+import { fileDownloadAndPreview } from '../../../../lib/methods/helpers';
+import { formatAttachmentUrl } from '../../../../lib/methods/helpers/formatAttachmentUrl';
+import openLink from '../../../../lib/methods/helpers/openLink';
+import { TSupportedThemes, useTheme } from '../../../../theme';
+import sharedStyles from '../../../../views/Styles';
+import RCActivityIndicator from '../../../ActivityIndicator';
+import Markdown from '../../../markdown';
 import Attachments from './Attachments';
-import MessageContext from './Context';
-import Touchable from './Touchable';
-import messageStyles from './styles';
+import MessageContext from '../../Context';
+import Touchable from '../../Touchable';
+import messageStyles from '../../styles';
 
 const styles = StyleSheet.create({
 	button: {
@@ -243,8 +243,7 @@ const Reply = React.memo(
 						}
 					]}
 					background={Touchable.Ripple(themes[theme].surfaceNeutral)}
-					disabled={!!(loading || attachment.message_link)}
-				>
+					disabled={!!(loading || attachment.message_link)}>
 					<View style={styles.attachmentContainer}>
 						<Title attachment={attachment} timeFormat={timeFormat} theme={theme} />
 						<Description attachment={attachment} getCustomEmoji={getCustomEmoji} theme={theme} />
@@ -264,8 +263,7 @@ const Reply = React.memo(
 									style={[
 										styles.backdrop,
 										{ backgroundColor: themes[theme].surfaceNeutral, opacity: themes[theme].attachmentLoadingOpacity }
-									]}
-								></View>
+									]}></View>
 								<RCActivityIndicator />
 							</View>
 						) : null}
