+++ app/containers/message/Components/Attachments/index.ts
@@ -1,0 +1,3 @@
+import Attachments from './Attachments';
+
+export default Attachments;
